make the conditions for converting weights to NZ format consistent
